Refactor: Move TV show episode entities to separate files
