+++ domain/src/main/java/com/london/domain/entity/tvshowdetails/episode/EpisodeCrewMemberEntity.kt
@@ -1,0 +1,15 @@
+package com.london.domain.entity.tvshowdetails.episode
+
+data class EpisodeCrewMemberEntity(
+    val job: String,
+    val department: String,
+    val creditId: String,
+    val adult: Boolean,
+    val gender: Int,
+    val id: Int,
+    val knownForDepartment: String,
+    val name: String,
+    val originalName: String,
+    val popularity: Double,
+    val profilePath: String?
+)
