+++ domain/src/main/java/com/london/domain/entity/tvshowdetails/episode/TvShowEpisodesEntity.kt
@@ -1,0 +1,7 @@
+package com.london.domain.entity.tvshowdetails.episode
+
+data class TvShowEpisodesEntity(
+    val id: String,
+    val airDate: String?,
+    val episodes: List<TvShowEpisodeBySeasonEntity>
+)
